Implment Maybe as classes plus helper functions
Here is another take on Maybe. This is a first pass at trying to implement this as I described in my comments to #13.

Not a huge fan of using classes, but its the only way I know of to define new _verifiable_ types in vanilla js.

Note that the design abstracts the class interface away from the user. Typical usage should not require importing the classes themselves unless you need to do type checking (for example type signatures in TypeScript).