Hide map chain value with non-enumerable props
This uses Object.create to define properties as non-enumerable, non-configurable and non-writable. The only exposed enumerable property is the purely informataive contents. Additionally instead of extending off Object.prototype, we extend off prototype of an empty Maybe function which makes justs and nothing both instances of a shared Maybe type."